# be/openapi.yaml
openapi: 3.0.0
info:
  title: fp1-Hotel Booking API (TypeScript)
  version: 1.0.0
  description: Typed API documentation for Hotel Booking System
servers:
  - url: http://localhost:3000
    description: Local Development Server

paths:
  # --- Authentication ---
  /auth/register:
    post:
      summary: Register a new user
      operationId: registerUser
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string }
                role: { $ref: '#/components/schemas/UserRole' }
                phone: { type: string }
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  statusCode:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/User'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Login and get JWT token
      operationId: loginUser
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  statusCode:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      token: { type: string }
                      user: { $ref: '#/components/schemas/User' }
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'



  # --- Customer ---
  /hotels/search:
    get:
      summary: Search Available Hotels
      tags: [Customer]
      security:
        - bearerAuth: []
      description: Search for hotels by city/name with availability, filtering, and sorting.
      operationId: searchAvailableHotels
      parameters:
        - name: query
          in: query
          description: Search term for City, Hotel Name, or Address (e.g., "Mumbai")
          required: true
          schema:
            type: string
            example: "Mumbai"
        - name: checkIn
          in: query
          description: Check-in date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
            example: "2024-12-01"
        - name: checkOut
          in: query
          description: Check-out date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
            example: "2024-12-05"
        - name: guests
          in: query
          description: Number of guests (to filter room capacity)
          required: false
          schema:
            type: integer
            default: 1
        - name: sortBy
          in: query
          description: Sort order for results
          required: false
          schema:
            type: string
            enum: [price_low, price_high, rating]
            default: "price_low"
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Successful search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  statusCode:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          description: Hotel ID
                        name:
                          type: string
                          description: Name of the hotel
                        city:
                          type: string
                        address:
                          type: string
                        description:
                          type: string
                        avg_rating:
                          type: number
                          format: float
                          example: 4.5
                        starting_price:
                          type: number
                          format: float
                          description: Lowest price found among room types
                        hotel_amenities:
                          type: string
                          description: Comma-separated list of hotel amenities (from GROUP_CONCAT)
                          example: "wifi,pool,parking"
                        manager_names:
                          type: array
                          items:
                            type: string
                          description: List of active managers (from JSON_ARRAYAGG)
                        rooms_details:
                          type: array
                          description: Nested list of rooms with their specific availability
                          items:
                            type: object
                            properties:
                              room_name:
                                type: string
                              price:
                                type: number
                              available_rooms:
                                type: integer
                                description: Calculated availability (Total - Booked)
                              amenities:
                                type: array
                                items:
                                  type: string
        "400":
          description: Bad Request (Invalid dates or missing parameters)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
        "401":
          description: Unauthorized
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string

  /hotels/{id}:
    get:
      summary: Get Hotel Details with Room Types
      operationId: getHotelById
      description: Returns full hotel details along with embedded room types and their availability.
      tags: [Customer]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200:
          description: Hotel details with rooms
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  statusCode:
                    type: integer
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Hotel'
                      - type: object
                        properties:
                          rooms:
                            type: array
                            items:
                              allOf:
                                - $ref: '#/components/schemas/RoomType'
                                - type: object
                                  properties:
                                    available:
                                      type: integer
                                      description: Calculated availability for the dates
        404:
          description: Hotel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a Hotel
      security:
        - bearerAuth: []
      operationId: deleteHotel
      tags: [Manager]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200:
          description: Hotel Deleted
          content:
             application/json:
               schema:
                 type: object
                 properties:
                   status: { type: boolean }
                   statusCode: { type: integer }
                   message: { type: string }
        403:
           description: Forbidden (Not a manager)
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorResponse'


  # --- Manager ---

  /hotels/my-hotels:
    get:
      summary: Get Hotels Managed by Current User
      security:
        - bearerAuth: []
      operationId: getMyHotels
      tags: [Manager]
      responses:
        200:
          description: List of hotels
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: boolean }
                  statusCode: { type: integer }
                  message: { type: string }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Hotel' # Now fully matching the Hotel DTO
        401:
           description: Unauthorized
  /hotels/create:
    post:
      summary: Create or Update a Hotel (Upsert)
      security:
        - bearerAuth: []
      operationId: createHotel
      tags: [Manager]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hotel'
      responses:
        201:
          description: Hotel Created
          content:
            application/json:
              schema:
                type: object
                properties:
                   status: { type: boolean }
                   message: { type: string }
                   data: { $ref: '#/components/schemas/Hotel' }
  
  /hotels/room-types:
    post:
      summary: Create or Update a Room Type (Upsert)
      security:
        - bearerAuth: []
      operationId: createRoomType
      tags: [Manager]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomType'
      responses:
        201:
          description: Room Type Created
          content:
            application/json:
              schema:
                type: object
                properties:
                   status: { type: boolean }
                   message: { type: string }
                   data: { $ref: '#/components/schemas/RoomType' }


  /hotels/room-types/{id}:
    delete:
      summary: Delete a Room Type
      security:
        - bearerAuth: []
      operationId: deleteRoomType
      tags: [Manager]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200:
          description: Room Type Deleted
          content:
             application/json:
               schema:
                 type: object
                 properties:
                   status: { type: boolean }
                   statusCode: { type: integer }
                   message: { type: string }
        403:
           description: Forbidden (Not a manager)
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorResponse'



  /hotels/manager/bookings:
    get:
      summary: Get All Bookings for a Hotel (Manager Only)
      security:
        - bearerAuth: []
      operationId: getManagerBookings
      tags: [Manager]
      responses:
        200:
          description: List of bookings with rich details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: boolean }
                  statusCode: { type: integer }
                  message: { type: string }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        hotelName: { type: string }
                        status: { type: string }
                        checkIn: { type: string, format: date-time }
                        checkOut: { type: string, format: date-time }
                        totalPrice: { type: number }
                        guestName: { type: string }
                        guestEmail: { type: string }
                        guestPhone: { type: string }
                        guests: 
                          type: array
                          items:
                            type: object
                            properties:
                              name: { type: string }
                              phone: { type: string }
                              email: { type: string }
                        rooms: { type: string, description: "Summary string of rooms" }
        403:
          description: Forbidden (Not a manager)


  # --- Bookings ---
  /bookings/intent:
    post:
      summary: Step 1 - Calculate Price & Initiate Payment
      security:
        - bearerAuth: []
      operationId: createBookingIntent
      description: |
        Calculates total price based on formatting, checks availability/inventory reduction, 
        and returns a Stripe Payment Intent client secret.
      tags: [Bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hotelId, roomTypeId, checkIn, checkOut, guests, paymentMode]
              properties:
                hotelId: { type: string }
                roomTypeId: { type: string }
                checkIn: { type: string, format: date }
                checkOut: { type: string, format: date }
                guests: { type: integer }
                paymentMode: { $ref: '#/components/schemas/PaymentMode' }
                guestDetails: # Required if PaymentMode is PAY_AT_HOTEL (can be single or array, usually array now)
                   type: array
                   items:
                     type: object
                     properties:
                       name: { type: string }
                       email: { type: string }
                       phone: { type: string }
      responses:
        200:
          description: Payment Intent Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  statusCode:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      paymentIntentId: { type: string }
                      clientSecret: { type: string }
                      totalPrice: { type: number }
                      currency: { type: string, example: "INR" }
        400:
          description: Validation error or room not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /bookings/confirm:
    post:
      summary: Step 2 - Finalize Booking
      security:
        - bearerAuth: []
      operationId: confirmBooking
      description: Called after successful payment on the client side to persist the booking in the DB.
      tags: [Bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, paymentIntentId]
              properties:
                bookingId: { type: string }
                paymentIntentId: { type: string }
                guestDetails: # Optional if we already have it from user profile, but good to confirm
                   type: array
                   items: 
                     type: object
                     properties:
                       name: { type: string }
                       email: { type: string }
                       phone: { type: string }
      responses:
        201:
          description: Booking Confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  statusCode:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Booking'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /bookings/my-bookings:
    get:
      summary: Get user's booking history
      security:
        - bearerAuth: []
      operationId: getUserBookings
      tags: [Bookings]
      responses:
        200:
          description: List of user bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  statusCode:
                    type: integer
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'



components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          example: false
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Error message details"
      
    UserRole:
      type: string
      enum: ['CUSTOMER', 'HOTEL_MANAGER', 'ADMIN']
    
    PaymentMode:
      type: string
      enum: ['ONLINE', 'PAY_AT_HOTEL']
    
    BookingStatus:
      type: string
      enum: ['PENDING_PAYMENT', 'CONFIRMED', 'CANCELLED', 'COMPLETED']

    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/UserRole' }
        phone: { type: string }
      required: [id, name, email, role]

    Amenity:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        scopeId: { type: string, nullable: true }
      required: [id, name]

    Hotel:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        city: { type: string }
        address: { type: string }
        description: { type: string, nullable: true }
        rating: { type: number }
        lowestPrice: { type: number, nullable: true }
        images: { type: array, items: { type: string } }
        amenities: { type: array, items: { type: string } }
        rooms: 
          type: array
          items: 
            $ref: '#/components/schemas/RoomType'
      required: [id, name, city, address, rating, images, amenities]

    RoomType:
      type: object
      properties:
        id: { type: string }
        hotelId: { type: string }
        name: { type: string }
        price: { type: number }
        capacity: { type: integer }
        totalInventory: { type: integer }
        images: { type: array, items: { type: string } }
        amenities: { type: array, items: { type: string } }
      required: [id, hotelId, name, price, capacity, totalInventory, images, amenities]

    Booking:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        hotelId: { type: string }
        roomTypeId: { type: string }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        guestDetails:
          type: object
          properties:
            name: { type: string }
            email: { type: string, format: email }
            phone: { type: string }
          required: [name, email, phone]
        totalPrice: { type: number }
        paymentMode: { $ref: '#/components/schemas/PaymentMode' }
        status: { $ref: '#/components/schemas/BookingStatus' }
        transactionId: { type: string }
      required: [id, userId, hotelId, roomTypeId, checkIn, checkOut, guestDetails, totalPrice, paymentMode, status]